+++ chaps/session_mock.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2012 The Chromium OS Authors. All rights reserved.
+// Copyright 2012 The ChromiumOS Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -66,7 +66,7 @@ class SessionMock : public Session {
                      int*,
                      int*));
   MOCK_METHOD1(SeedRandom, CK_RV(const std::string&));
-  MOCK_METHOD2(GenerateRandom, CK_RV(int, std::string*));
+  MOCK_METHOD2(GenerateRandom, CK_RV(size_t, std::string*));
   MOCK_METHOD0(IsPrivateLoaded, bool());
 };
